Move classes in go_router.dart into separate libraries
