+++ packages/go_router/lib/src/misc/extensions.dart
@@ -1,0 +1,59 @@
+import 'package:flutter/widgets.dart';
+
+import '../router.dart';
+
+/// Dart extension to add navigation function to a BuildContext object, e.g.
+/// context.go('/');
+extension GoRouterHelper on BuildContext {
+  /// Get a location from route name and parameters.
+  String namedLocation(
+      String name, {
+        Map<String, String> params = const <String, String>{},
+        Map<String, String> queryParams = const <String, String>{},
+      }) =>
+      GoRouter.of(this)
+          .namedLocation(name, params: params, queryParams: queryParams);
+
+  /// Navigate to a location.
+  void go(String location, {Object? extra}) =>
+      GoRouter.of(this).go(location, extra: extra);
+
+  /// Navigate to a named route.
+  void goNamed(
+      String name, {
+        Map<String, String> params = const <String, String>{},
+        Map<String, String> queryParams = const <String, String>{},
+        Object? extra,
+      }) =>
+      GoRouter.of(this).goNamed(
+        name,
+        params: params,
+        queryParams: queryParams,
+        extra: extra,
+      );
+
+  /// Push a location onto the page stack.
+  void push(String location, {Object? extra}) =>
+      GoRouter.of(this).push(location, extra: extra);
+
+  /// Navigate to a named route onto the page stack.
+  void pushNamed(
+      String name, {
+        Map<String, String> params = const <String, String>{},
+        Map<String, String> queryParams = const <String, String>{},
+        Object? extra,
+      }) =>
+      GoRouter.of(this).pushNamed(
+        name,
+        params: params,
+        queryParams: queryParams,
+        extra: extra,
+      );
+
+  /// Returns `true` if there is more than 1 page on the stack.
+  bool canPop() => GoRouter.of(this).canPop();
+
+  /// Pop the top page off the Navigator's page stack by calling
+  /// [Navigator.pop].
+  void pop() => GoRouter.of(this).pop();
+}
