+++ packages/go_router/lib/src/pages/material.dart
@@ -5,7 +5,8 @@
 // ignore_for_file: diagnostic_describe_all_properties
 
 import 'package:flutter/material.dart';
-import '../../go_router.dart';
+
+import '../misc/extensions.dart';
 
 /// Checks for MaterialApp in the widget tree.
 bool isMaterialApp(Element elem) =>
